refactor(coord): make coord_radial the base, coord_polar its alias
coord_radial now owns the polar logic directly (parent bodies inlined,
theta breaks computed once) and coord_polar is an @alias subclass.
coord_polar() now shows theta tick labels by default.
