+++ plotnine/coords/__init__.py
@@ -5,8 +5,7 @@ Coordinates
 from .coord_cartesian import coord_cartesian
 from .coord_fixed import coord_equal, coord_fixed
 from .coord_flip import coord_flip
-from .coord_polar import coord_polar
-from .coord_radial import coord_radial
+from .coord_radial import coord_polar, coord_radial
 from .coord_trans import coord_trans
 
 __all__ = (
